+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## [2.6.0]
+- Synchroniser les infos sur l'accessibilité des treks sur Apidae [#120](https://github.com/GeotrekCE/Sitourisme/issues/120)
 - Faire remonter les traductions des réseaux de randonnée (balisage) [#119](https://github.com/GeotrekCE/Sitourisme/issues/119)
 - Ajouter l'information Animaux interdits si passage zone Coeur de Parc [#126](https://github.com/GeotrekCE/Sitourisme/issues/126)
 - Trek > Synchroniser le champ Matériel dans Prestations > ComplémentAccueil [#128](https://github.com/GeotrekCE/Sitourisme/issues/128)